% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.provide.R
\name{col_palette}
\alias{col_palette}
\title{Returns a named list of all colours; or a data frame containing names, hex codes and RGB values.}
\usage{
col_palette(return.df = FALSE)
}
\arguments{
\item{return.df}{logical. Set to TRUE if a data frame should be returned. Defaults to FALSE.}
}
\description{
Returns a named list of all colours; or a data frame containing names, hex codes and RGB values.
}
