% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.provide.R
\name{get_col}
\alias{get_col}
\title{Returns the colour value given a valid colour name.}
\usage{
get_col(cname = NULL, alpha = 1, return.rgb = FALSE)
}
\arguments{
\item{cname}{Colour name. See col_names for a list of valid names. Returns default blue if no name is specified.}

\item{alpha}{Optional transparency value. Set to 1 by default.}

\item{return.rgb}{logical. Use when RGB values should be returned instead of hex value.}
}
\description{
Returns the colour value given a valid colour name.
}
