% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.provide.R
\name{show_colours}
\alias{show_colours}
\title{Shows the UAS Mittweida colour set}
\usage{
show_colours()
}
\description{
A simple graphical output listing all of the colours provided by this package. Please note that
UAS red and green colours are simple RGB complementaries to
the official blues and are not officially part of the corporate design.
However these reds and greens might be used for nonofficial purposes in cases
where showing categorical data requires more contrast and colour levels than supported by
the three CD blues, which would be more appropriate anyway.
}
